+++ tools/rpkg/CMakeLists.txt
@@ -22,6 +22,16 @@ include_directories(${R_INCLUDE_DIR})
 include_directories(../../)
 
 # this is used for clang-tidy checks and IDE support
-add_library(duckdb_r_shlib SHARED src/duckdbr.cpp)
+add_library(
+  duckdb_r_shlib SHARED
+  src/duckdbr.cpp
+  src/database.cpp
+  src/connection.cpp
+  src/statement.cpp
+  src/register.cpp
+  src/scan.cpp
+  src/utils.cpp
+  src/altrep.cpp
+  src/types.cpp)
 target_link_libraries(duckdb_r_shlib duckdb ${DUCKDB_EXTRA_LINK_FLAGS}
                       ${R_LIBRARY_BASE})
